Moves printing to RegionImport. Makes code more Object Oriented.
