+++ src/main/scala/scala/tools/refactoring/implementations/oimports/Region.scala
@@ -1,7 +1,6 @@
 package scala.tools.refactoring
 package implementations.oimports
 
-import scala.reflect.internal.util.Position
 import scala.reflect.internal.util.RangePosition
 import scala.reflect.internal.util.SourceFile
 import scala.tools.nsc.Global
@@ -10,11 +9,10 @@ import scala.tools.refactoring.common.TextChange
 import scala.util.Properties
 
 import sourcegen.Formatting
-import scala.reflect.internal.Chars
 
 case class Region private (imports: List[Global#Import], owner: Global#Symbol, from: Int,
-    to: Int, source: SourceFile, indentation: String, printImport: Global#Import => String,
-    printImportWithComment: Global#Import => String, printWhenEmpty: String) {
+    to: Int, source: SourceFile, indentation: String,
+    formatting: Formatting, printWhenEmpty: String) {
   def transform(transformation: List[Global#Import] => List[Global#Import]): Region =
     copy(imports = transformation(imports))
 
@@ -41,14 +39,14 @@ case class Region private (imports: List[Global#Import], owner: Global#Symbol, f
     val text = imports.zipWithIndex.foldLeft("") { (acc, imp) =>
       def isLast(idx: Int) = idx == imports.size - 1
       imp match {
-        case (imp, 0) if isLast(0) =>
-          acc + printImportWithComment(imp)
-        case (imp, 0) =>
-          acc + printImportWithComment(imp) + Properties.lineSeparator
-        case (imp, idx) if isLast(idx) =>
-          acc + indentation + printImportWithComment(imp)
-        case (imp, _) =>
-          acc + indentation + printImportWithComment(imp) + Properties.lineSeparator
+        case (imp: TreeToolbox[_]#RegionImport, 0) if isLast(0) =>
+          acc + imp.printWithComment(formatting)
+        case (imp: TreeToolbox[_]#RegionImport, 0) =>
+          acc + imp.printWithComment(formatting) + Properties.lineSeparator
+        case (imp: TreeToolbox[_]#RegionImport, idx) if isLast(idx) =>
+          acc + indentation + imp.printWithComment(formatting)
+        case (imp: TreeToolbox[_]#RegionImport, _) =>
+          acc + indentation + imp.printWithComment(formatting) + Properties.lineSeparator
       }
     }
     TextChange(source, from, to, text)
@@ -56,40 +54,6 @@ case class Region private (imports: List[Global#Import], owner: Global#Symbol, f
 }
 
 object Region {
-  private def indentation(imp: Global#Import): String = {
-    val sourceFile = imp.pos.source
-    sourceFile.lineToString(sourceFile.offsetToLine(imp.pos.start)).takeWhile { _.isWhitespace }
-  }
-
-  private def decodedName(keywords: Set[String])(name: Global#Name) = {
-    def addBackquotes(str: String) = {
-      val (ident, op) =
-        if (Chars.isIdentifierStart(str.head))
-          str.span(Chars.isIdentifierPart)
-        else
-          ("", str)
-      val needsBackticks =
-        if (op.isEmpty)
-          keywords(name.toTermName.decoded) && name.toTermName.decoded != "_"
-        else if (!ident.isEmpty && ident.last != '_')
-          true
-        else
-          !op.tail.forall(Chars.isOperatorPart)
-      if (needsBackticks) s"`$str`" else str
-    }
-    addBackquotes(name.decoded.trim)
-  }
-
-  private def mkFromSelector(keywords: Set[String])(sel: Global#ImportSelector): String = {
-    val renameArrow = " => "
-    decodedName(keywords)(sel.name) + {
-      if (sel.rename != null && sel.name.decoded != sel.rename.decoded)
-        renameArrow + decodedName(keywords)(sel.rename)
-      else
-        ""
-    }
-  }
-
   private def scanForComments[G <: Global](global: G)(source: SourceFile): List[RangePosition] = {
     val scanners = new TreeToolboxScanners(global)
     val commentScanner = new scanners.CommentScanner(source)
@@ -97,93 +61,17 @@ object Region {
     commentScanner.comments
   }
 
-  private def cutPrefixSuffix(imp: Global#Import, source: SourceFile): (String, List[String]) = {
-    val printedImport = source.content.slice(imp.pos.start, imp.pos.end).mkString
-    val prefixPatternWithCommentInside = """import\s+(((\/\*.*\*\/)*((\w|\d|_|-)+|(\`.*\`)+)(\/\*.*\*\/)*)\.)+(\/\*.*\*\/)*""".r
-    val prefix = prefixPatternWithCommentInside.findFirstIn(printedImport).get
-    def toNameRename(printedSelectors: String): List[String] = {
-      val unwrapFromBraces = (if (printedSelectors.startsWith("{"))
-        printedSelectors.drop(1).dropRight(1)
-      else printedSelectors).split(",").filter { _ != "" }.map { _.trim }
-      unwrapFromBraces.toList
-    }
-    val rawSelectors = printedImport.substring(prefix.length).trim
-    (prefix, toNameRename(rawSelectors))
-  }
-
-  private def selectorToSuffix(suffices: List[String], sel: Global#ImportSelector): Option[String] = suffices.find { s =>
-    val name = sel.name.decoded
-    val backtickedName = "`" + name + "`"
-    val rename = if (sel.rename != null) sel.rename.decoded else ""
-    val backtickedRename = "`" + rename + "`"
-    def isCorrect(s: String): Boolean = {
-      val renameArrowOrNothingLeft = """(\s*=>\s*)?""".r
-      renameArrowOrNothingLeft.findAllIn(s).nonEmpty
-    }
-    val found = if (s.startsWith(name)) {
-      if (s.endsWith(backtickedRename)) {
-        isCorrect(s.drop(name.length).dropRight(backtickedRename.length))
-      } else if (s.endsWith(rename)) {
-        isCorrect(s.drop(name.length).dropRight(rename.length))
-      } else false
-    } else if (s.startsWith(backtickedName)) {
-      if (s.endsWith(backtickedRename)) {
-        isCorrect(s.drop(backtickedName.length).dropRight(backtickedRename.length))
-      } else if (s.endsWith(rename)) {
-        isCorrect(s.drop(backtickedName.length).dropRight(rename.length))
-      } else false
-    } else false
-    found
-  }
-
-  private def findUpNeighborComment(impPos: Position, comments: List[RangePosition], source: SourceFile): Option[RangePosition] = impPos match {
-    case rangePos: RangePosition =>
-      val beginningOfImportLine = source.lineToOffset(source.offsetToLine(rangePos.start))
-      comments.find { comment =>
-        comment.end == beginningOfImportLine
-      }
-    case _ => None
-  }
-
-  private def findUpNeighborCommentText(impPos: Position, comments: List[RangePosition], source: SourceFile): Option[String] =
-    findUpNeighborComment(impPos, comments, source).map { comment =>
-      source.content.slice(comment.start, comment.end).mkString
-    }
-
-  private def wrapInBraces(selectors: String, rawSelectors: List[Global#ImportSelector], formatting: Formatting): String =
-    if (rawSelectors.length > 1 || rawSelectors.exists { sel =>
-      sel.rename != null && sel.name.decoded != sel.rename.decoded
-    })
-      "{" + formatting.spacingAroundMultipleImports + selectors + formatting.spacingAroundMultipleImports + "}"
-    else
-      selectors
-
   def apply[G <: Global, T <: TreeToolbox[G]](treeToolbox: T)(imports: List[treeToolbox.global.Import], owner: treeToolbox.global.Symbol, formatting: Formatting): Region = {
     require(imports.nonEmpty, "List of imports must not be empty.")
-    import treeToolbox.global.nme
     val source = imports.head.pos.source
     val comments = scanForComments(treeToolbox.global)(source)
-    def printImport(imp: Global#Import): String = {
-      val (prefix, suffices) = cutPrefixSuffix(imp, source)
-      val keywords = nme.keywords.map { _.decoded }
-      val suffix = imp.selectors.collect {
-        case sel => selectorToSuffix(suffices, sel).orElse(Option(mkFromSelector(keywords)(sel)))
-      }.filter(_.nonEmpty).map(_.get)
-      prefix + wrapInBraces(suffix.mkString(", "), imp.selectors, formatting)
-    }
-    val indent = indentation(imports.head)
-    def printImportWithComment(imp: Global#Import): String = {
-      val printedImport = printImport(imp)
-      findUpNeighborCommentText(imp.pos, comments, source).map { comment =>
-        comment + indent + printedImport
-      }.getOrElse(printedImport)
-    }
-    val regionStartPos = findUpNeighborComment(imports.head.pos, comments, source).getOrElse(imports.head.pos)
-    Region(toRegionImports[G, T](treeToolbox)(imports, owner), owner, regionStartPos.start, imports.last.pos.end, source, indent,
-        printImport, printImportWithComment, "")
+    val regionImports = toRegionImports[G, T](treeToolbox)(imports, owner, comments)
+    val regionStartPos = regionImports.head.positionWithComment.start
+    val indentation = regionImports.head.indentation
+    Region(regionImports, owner, regionStartPos, imports.last.pos.end, source, indentation, formatting, "")
   }
 
-  private def toRegionImports[G <: Global, T <: TreeToolbox[G]](ttb: T)(imports: List[ttb.global.Import], owner: ttb.global.Symbol) = {
-    imports.map { i => new ttb.RegionImport(owner, i)() }
+  private def toRegionImports[G <: Global, T <: TreeToolbox[G]](ttb: T)(imports: List[ttb.global.Import], owner: ttb.global.Symbol, comments: List[RangePosition]) = {
+    imports.map { i => new ttb.RegionImport(owner, i, comments)() }
   }
 }
